Move FindOldestTransactionTime to `executor` package.
